Rename "product" variables to "app" variables
