+++ data/colors/themes/dark.ts
@@ -3,7 +3,7 @@ import darkComponentVars from '../vars/component_dark'
 import darkDeprecatedVars from '../vars/deprecated_dark'
 import darkGlobalVars from '../vars/global_dark'
 import darkMarketingVars from '../vars/marketing_dark'
-import darkProductVars from '../vars/product_dark'
+import darkAppVars from '../vars/app_dark'
 
 const scale = {
   black: '#010409',
@@ -63,4 +63,4 @@ const scale = {
   ]
 }
 
-export default merge(darkDeprecatedVars, darkProductVars, darkMarketingVars, darkComponentVars, darkGlobalVars, {scale})
+export default merge(darkDeprecatedVars, darkAppVars, darkMarketingVars, darkComponentVars, darkGlobalVars, {scale})
